Issue with getting module to read and load
