+++ app/views/generator/_form_stacked.html.erb
@@ -22,7 +22,7 @@
     <div class="row">
       <div id="generator-autocomplete-destination" class="columns small-12">
         <label>Destination
-          <input id="autocomplete-destination" placeholder="Destination" type="text" name="destination_point"></input>
+          <input id="autocomplete-destination" placeholder="Address, city, or state etc" type="text" name="destination_point"></input>
         </label>
       </div>
     </div>
